+++ network/underlay/noop.go
@@ -1,4 +1,4 @@
-package network
+package underlay
 
 import (
 	"github.com/ipfs/go-datastore"
@@ -6,14 +6,15 @@ import (
 
 	"github.com/onflow/flow-go/model/flow"
 	"github.com/onflow/flow-go/module"
+	"github.com/onflow/flow-go/network"
 	"github.com/onflow/flow-go/network/channels"
 )
 
 type NoopConduit struct{}
 
-var _ Conduit = (*NoopConduit)(nil)
+var _ network.Conduit = (*NoopConduit)(nil)
 
-func (n *NoopConduit) ReportMisbehavior(MisbehaviorReport) {}
+func (n *NoopConduit) ReportMisbehavior(network.MisbehaviorReport) {}
 
 func (n *NoopConduit) Publish(event interface{}, targetIDs ...flow.Identifier) error {
 	return nil
@@ -35,16 +36,16 @@ type NoopEngineRegister struct {
 	module.NoopComponent
 }
 
-func (n NoopEngineRegister) Register(channel channels.Channel, messageProcessor MessageProcessor) (Conduit, error) {
+func (n NoopEngineRegister) Register(channel channels.Channel, messageProcessor network.MessageProcessor) (network.Conduit, error) {
 	return &NoopConduit{}, nil
 }
 
-func (n NoopEngineRegister) RegisterBlobService(channel channels.Channel, store datastore.Batching, opts ...BlobServiceOption) (BlobService, error) {
+func (n NoopEngineRegister) RegisterBlobService(channel channels.Channel, store datastore.Batching, opts ...network.BlobServiceOption) (network.BlobService, error) {
 	return nil, nil
 }
 
-func (n NoopEngineRegister) RegisterPingService(pingProtocolID protocol.ID, pingInfoProvider PingInfoProvider) (PingService, error) {
+func (n NoopEngineRegister) RegisterPingService(pingProtocolID protocol.ID, pingInfoProvider network.PingInfoProvider) (network.PingService, error) {
 	return nil, nil
 }
 
-var _ EngineRegistry = (*NoopEngineRegister)(nil)
+var _ network.EngineRegistry = (*NoopEngineRegister)(nil)
